requirements.txt
meshcore
metpy
PyYAML
python-dotenv
black
loguru
pysqlite3
isort
croniter
yamlfix
mdformat
pylint
